% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_cols.R
\name{add_cols}
\alias{add_cols}
\title{cbind two data frames at arbitrary position}
\usage{
add_cols(x, y, where = 1)
}
\arguments{
\item{x, y}{Data frames to combine}

\item{where}{Position of \code{y} relative to \code{x}.
\code{where = 1} will insert \code{y} on the left,
\code{where = ncol(x) + 1} will insert \code{y} on the
right.}
}
\value{
A data frame
}
\description{
A more flexible variant of \code{\link[=cbind]{cbind()}}.
}
