% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_into.R
\name{insert_into}
\alias{insert_into}
\title{Insert_into
insert a dataframe / col into anther data.frame .}
\usage{
insert_into(x, y, where = 1)
}
\arguments{
\item{x}{a data.frame}

\item{y}{a data.frame}

\item{where}{an integer}
}
\value{
the combined data.frame
}
\description{
Insert_into
insert a dataframe / col into anther data.frame .
}
