% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_into.R
\name{insert_into}
\alias{insert_into}
\title{Insert a data.frame into another data.frame}
\usage{
insert_into(x, y, where = 1)
}
\arguments{
\item{x}{A \link[base]{data.frame}}

\item{y}{A \code{\link[base:data.frame]{base::data.frame()}}}

\item{where}{The place where data y will be inserted into data x}
}
\value{
combined data
}
\description{
Insert a data.frame into another data.frame
}
\examples{
x <- data.frame(a = "abc")
y <- data.frame(b = "def")
insert_into(x,y)
insert_into(x,y,where = 2)
}
\seealso{
\code{\link[=add_col]{add_col()}}

Other xyz: \code{\link{add_col}}
}
\concept{xyz}
